Python Projects/Database Fill/requirements.txt
Faker==5.0.0
mysql-connector-python==8.0.22
names==0.3.0
protobuf==3.14.0
python-dateutil==2.8.1
six==1.15.0
text-unidecode==1.3
